ANGIRA_VOICE/requirements.txt
# Google Gemini SDK (new unified SDK)
google-genai>=1.0.0

# Other dependencies
openai==1.3.0
python-dotenv==1.0.0
numpy>=1.24.3
scipy>=1.11.1
librosa>=0.10.0
soundfile>=0.12.1
pyaudio>=0.2.13
typing-extensions>=4.7.1
pydantic>=2.0.0
